Avoid using extern for a simple header/source separation
